# expression: implement vectorized evaluation for `builtinTimestamp1ArgSig`