--- a/expression/bench_test.go
+++ b/expression/bench_test.go
@@ -326,6 +326,18 @@ func (g *dataTimeStrGener) gen() interface{} {
 		year, month, day, hour, minute, second)
 }
 
+// timeStrGener is used to generate strings which are time format
+type timeStrGener struct{}
+
+func (g *timeStrGener) gen() interface{} {
+	year := rand.Intn(2200)
+	month := rand.Intn(10) + 1
+	day := rand.Intn(20) + 1
+
+	return fmt.Sprintf("%d-%d-%d",
+		year, month, day)
+}
+
 // dataStrGener is used to generate strings which are data format
 type dataStrGener struct{}