<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for ```builtinTimestamp1ArgSig```
To: #12101 

### What is changed and how it works?
```
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestamp1ArgSig-VecBuiltinFunc-4                 2416            498074 ns/op          147457 B/op       7168 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestamp1ArgSig-NonVecBuiltinFunc-4              2119            585147 ns/op          147456 B/op       7168 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestamp1ArgSig-VecBuiltinFunc#01-4              2797            436687 ns/op          147456 B/op       7168 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestamp1ArgSig-NonVecBuiltinFunc#01-4           2124            542507 ns/op          147456 B/op       7168 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestamp1ArgSig-VecBuiltinFunc#02-4               337           3399349 ns/op          569341 B/op      16130 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestamp1ArgSig-NonVecBuiltinFunc#02-4            252           4736968 ns/op          569340 B/op      16130 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestamp1ArgSig-VecBuiltinFunc#03-4               318           3641208 ns/op          475200 B/op      10725 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestamp1ArgSig-NonVecBuiltinFunc#03-4            235           4978121 ns/op          475200 B/op      10725 allocs/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test