@zz-jason 2019-09-24
The performance is really poor compared with other builtin functions :joy: